Changing datatype Real_t from float to double

+++ Source_Code_4D/XT_Constants.h
@@ -42,9 +42,9 @@
 	#define DEBUG_LOW
 #endif
 
-typedef float Real_t;
-#define MPI_TEMP_DATATYPE MPI_FLOAT
-#define MPI_OBJECT_DATATYPE MPI_FLOAT
+typedef double Real_t;
+#define MPI_TEMP_DATATYPE MPI_DOUBLE
+#define MPI_OBJECT_DATATYPE MPI_DOUBLE
 
 #define ZERO_SKIPPING
 #define INIT_SPARSE_ANGLES_FROM_FILE
